Reorganize native code
