+++ graphql-native/src/main/java/io/ballerina/stdlib/graphql/runtime/schema/TypeFinder.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.runtime.schema.tree;
+package io.ballerina.stdlib.graphql.runtime.schema;
 
 import io.ballerina.runtime.api.TypeTags;
 import io.ballerina.runtime.api.creators.ValueCreator;
@@ -29,8 +29,8 @@ import io.ballerina.runtime.api.types.ServiceType;
 import io.ballerina.runtime.api.types.TableType;
 import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.types.UnionType;
-import io.ballerina.stdlib.graphql.runtime.schema.SchemaType;
-import io.ballerina.stdlib.graphql.runtime.schema.TypeKind;
+import io.ballerina.stdlib.graphql.runtime.schema.types.SchemaType;
+import io.ballerina.stdlib.graphql.runtime.schema.types.TypeKind;
 
 import java.util.HashMap;
 import java.util.List;
@@ -38,9 +38,9 @@ import java.util.Map;
 
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.QUERY;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.SCHEMA_RECORD;
-import static io.ballerina.stdlib.graphql.runtime.schema.tree.Utils.getScalarTypeName;
-import static io.ballerina.stdlib.graphql.runtime.schema.tree.Utils.getTypeNameFromType;
-import static io.ballerina.stdlib.graphql.runtime.schema.tree.Utils.isEnum;
+import static io.ballerina.stdlib.graphql.runtime.schema.Utils.getScalarTypeName;
+import static io.ballerina.stdlib.graphql.runtime.schema.Utils.getTypeNameFromType;
+import static io.ballerina.stdlib.graphql.runtime.schema.Utils.isEnum;
 import static io.ballerina.stdlib.graphql.runtime.utils.ModuleUtils.getModule;
 import static io.ballerina.stdlib.graphql.runtime.utils.Utils.NOT_SUPPORTED_ERROR;
 import static io.ballerina.stdlib.graphql.runtime.utils.Utils.createError;
